+++ src/main/java/ch/bitagent/bitcoin/lib/wallet/ExtendedKey.java
@@ -5,6 +5,7 @@ import ch.bitagent.bitcoin.lib.ecc.Int;
 import ch.bitagent.bitcoin.lib.ecc.PrivateKey;
 import ch.bitagent.bitcoin.lib.ecc.S256Point;
 import ch.bitagent.bitcoin.lib.helper.Base58;
+import ch.bitagent.bitcoin.lib.helper.Bytes;
 import ch.bitagent.bitcoin.lib.helper.Hash;
 
 import java.util.Arrays;
@@ -49,7 +50,7 @@ public class ExtendedKey {
             throw new IllegalArgumentException("Invalid extendend key length");
         }
         this.prefix = Arrays.copyOfRange(decoded, 0, 4);
-        this.isKeyPrivate();
+        isKeyPrivate(this.prefix);
         this.depth = decoded[4];
         this.fingerprint = Arrays.copyOfRange(decoded, 5, 9);
         this.childNumber = Arrays.copyOfRange(decoded, 9, 13);
@@ -63,7 +64,7 @@ public class ExtendedKey {
         }
         this.chainCode = Arrays.copyOfRange(decoded, 13, 45);
         this.key = Arrays.copyOfRange(decoded, 45, decoded.length);
-        if (this.isKeyPrivate()) {
+        if (isKeyPrivate(this.prefix)) {
             var hexKey = Hex.parse(this.key);
             var one = Int.parse(1);
             if (hexKey.lt(one) || hexKey.gt(S256Point.N.sub(one))) {
@@ -89,7 +90,7 @@ public class ExtendedKey {
         if (this.prefix.length != 4) {
             throw new IllegalArgumentException("invalid prefix");
         }
-        this.isKeyPrivate();
+        isKeyPrivate(this.prefix);
         this.depth = depth;
         this.fingerprint = fingerprint;
         if (this.fingerprint.length != 4) {
@@ -109,21 +110,21 @@ public class ExtendedKey {
         }
     }
 
-    public boolean isKeyPrivate() {
-        if (Arrays.equals(PREFIX_XPRV.toBytes(), this.prefix)) {
+    public static boolean isKeyPrivate(byte[] prefix) {
+        if (Arrays.equals(PREFIX_XPRV.toBytes(), prefix)) {
             return true;
-        } else if (Arrays.equals(PREFIX_XPUB.toBytes(), this.prefix)) {
+        } else if (Arrays.equals(PREFIX_XPUB.toBytes(), prefix)) {
             return false;
-        } else if (Arrays.equals(PREFIX_YPRV.toBytes(), this.prefix)) {
+        } else if (Arrays.equals(PREFIX_YPRV.toBytes(), prefix)) {
             return true;
-        } else if (Arrays.equals(PREFIX_YPUB.toBytes(), this.prefix)) {
+        } else if (Arrays.equals(PREFIX_YPUB.toBytes(), prefix)) {
             return false;
-        } else if (Arrays.equals(PREFIX_ZPRV.toBytes(), this.prefix)) {
+        } else if (Arrays.equals(PREFIX_ZPRV.toBytes(), prefix)) {
             return true;
-        } else if (Arrays.equals(PREFIX_ZPUB.toBytes(), this.prefix)) {
+        } else if (Arrays.equals(PREFIX_ZPUB.toBytes(), prefix)) {
             return false;
         } else {
-            throw new IllegalArgumentException(String.format("Unknown prefix %s", Hex.parse(this.prefix)));
+            throw new IllegalArgumentException(String.format("Unknown prefix %s", Hex.parse(prefix)));
         }
     }
 
@@ -139,6 +140,27 @@ public class ExtendedKey {
         }
     }
 
+    public String serialize(boolean neutral) {
+        byte[] xkey;
+        if (isKeyPrivate(this.prefix) && neutral) {
+            xkey = getPrefixNeutral();
+        } else {
+            xkey = this.prefix;
+        }
+        xkey = Bytes.add(xkey, new byte[]{(byte) this.depth});
+        xkey = Bytes.add(xkey, this.fingerprint);
+        xkey = Bytes.add(xkey, this.childNumber);
+        xkey = Bytes.add(xkey, this.chainCode);
+        if (isKeyPrivate(this.prefix) && neutral) {
+            xkey = Bytes.add(xkey, PrivateKey.parse(this.key).getPoint().sec(true));
+        } else {
+            xkey = Bytes.add(xkey, this.key);
+        }
+        var hashedXprv = Hash.hash256(xkey);
+        xkey = Bytes.add(xkey, Arrays.copyOfRange(hashedXprv, 0, 4)); // Checksum
+        return Base58.encode(xkey);
+    }
+
     public ExtendedKey derive(int index) {
         return this.derive(index, false, false);
     }
@@ -148,7 +170,7 @@ public class ExtendedKey {
         if (harden) {
             indexInt = indexInt.add(HARDENED_INDEX);
         }
-        if (this.isKeyPrivate()) {
+        if (isKeyPrivate(this.prefix)) {
             var derivedDepth = this.depth + 1;
             var privateKey = PrivateKey.parse(this.key);
             var publicKeyPoint = privateKey.getPoint();